+++ crates/eww/src/widgets/widget_definitions.rs
@@ -3,7 +3,7 @@ use super::{build_widget::BuilderArgs, circular_progressbar::*, run_command, tra
 use crate::{
     def_widget, enum_parse, error_handling_ctx,
     util::{self, list_difference},
-    widgets::build_widget::build_gtk_widget,
+    widgets::{build_widget::build_gtk_widget, systray},
 };
 use anyhow::{anyhow, Context, Result};
 use codespan_reporting::diagnostic::Severity;
@@ -82,6 +82,7 @@ pub const BUILTIN_WIDGET_NAMES: &[&str] = &[
     WIDGET_NAME_SCROLL,
     WIDGET_NAME_OVERLAY,
     WIDGET_NAME_STACK,
+    WIDGET_NAME_SYSTRAY,
 ];
 
 /// widget definitions
@@ -111,6 +112,7 @@ pub(super) fn widget_use_to_gtk_widget(bargs: &mut BuilderArgs) -> Result<gtk::W
         WIDGET_NAME_SCROLL => build_gtk_scrolledwindow(bargs)?.upcast(),
         WIDGET_NAME_OVERLAY => build_gtk_overlay(bargs)?.upcast(),
         WIDGET_NAME_STACK => build_gtk_stack(bargs)?.upcast(),
+        WIDGET_NAME_SYSTRAY => build_systray(bargs)?.upcast(),
         _ => {
             return Err(DiagError(gen_diagnostic! {
                 msg = format!("referenced unknown widget `{}`", bargs.widget_use.name),
@@ -1133,6 +1135,33 @@ fn build_graph(bargs: &mut BuilderArgs) -> Result<super::graph::Graph> {
     Ok(w)
 }
 
+const WIDGET_NAME_SYSTRAY: &str = "systray";
+/// @widget systray
+/// @desc Tray for system notifier icons
+fn build_systray(bargs: &mut BuilderArgs) -> Result<gtk::MenuBar> {
+    let gtk_widget = gtk::MenuBar::new();
+    let props = Rc::new(systray::Props::new());
+    let props_clone = props.clone();
+
+    // copies for def_widget
+    def_widget!(bargs, _g, gtk_widget, {
+        // @prop icon-size - size of icons in the tray
+        prop(icon_size: as_i32) {
+            if icon_size <= 0 {
+                log::warn!("Icon size is not a positive number");
+            } else {
+                props.icon_size(icon_size);
+            }
+        },
+        // @prop pack-direction - how to arrange tray items
+        prop(pack_direction: as_string) { gtk_widget.set_pack_direction(parse_packdirection(&pack_direction)?); },
+    });
+
+    systray::spawn_systray(&gtk_widget, &props_clone);
+
+    Ok(gtk_widget)
+}
+
 /// @var orientation - "vertical", "v", "horizontal", "h"
 fn parse_orientation(o: &str) -> Result<gtk::Orientation> {
     enum_parse! { "orientation", o,
@@ -1207,6 +1236,16 @@ fn parse_gravity(g: &str) -> Result<gtk::pango::Gravity> {
         "west" => gtk::pango::Gravity::West,
         "north" => gtk::pango::Gravity::North,
         "auto" => gtk::pango::Gravity::Auto,
+    }
+}
+
+/// @var pack-direction - "right", "ltr", "left", "rtl", "down", "ttb", "up", "btt"
+fn parse_packdirection(o: &str) -> Result<gtk::PackDirection> {
+    enum_parse! { "packdirection", o,
+        "right" | "ltr" => gtk::PackDirection::Ltr,
+        "left" | "rtl" => gtk::PackDirection::Rtl,
+        "down" | "ttb" => gtk::PackDirection::Ttb,
+        "up" | "btt" => gtk::PackDirection::Btt,
     }
 }