+++ crates/simplexpr/src/eval.rs
@@ -313,6 +313,13 @@ impl SimplExpr {
 
 fn call_expr_function(name: &str, args: Vec<DynVal>) -> Result<DynVal, EvalError> {
     match name {
+        "get_env" => match args.as_slice() {
+            [var_name] => {
+                let var = std::env::var(var_name.as_string()?).unwrap_or_default();
+                Ok(DynVal::from(var))
+            }
+            _ => Err(EvalError::WrongArgCount(name.to_string())),
+        },
         "round" => match args.as_slice() {
             [num, digits] => {
                 let num = num.as_f64()?;